aintelliscrape/requirements.txt
aws-wsgi
pandas
python-docx
dash
